Moved CompleteInjection to parse.go
